refactor: Separate export code from CLI code
Export code now in the exporter folder and CLI code in the client.py file.
