+++ osfio-export-tool/src/clitool/__init__.py
@@ -1,12 +1,7 @@
 from .client import (
-    cli, call_api, get_project_data,
-    explore_file_tree, explore_wikis
+    cli
 )
 
 __all__ = [
-    'cli',
-    'call_api',
-    'get_project_data',
-    'explore_file_tree',
-    'explore_wikis'
+    'cli'
 ]
